feat: implementa metodos para repository

+++ gestao-escolar/src/main/java/com/weg/gestao_escolar/domain/repository/AulaRepository.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface AulaRepository {
     Aula save(Aula aula) throws SQLException;
-    Aula findById(Long id) throws SQLException;
+    Aula findById(int id) throws SQLException;
     List<Aula> findAll() throws SQLException;
     void update(Aula aula) throws SQLException;
-    void delete(Long id) throws SQLException;
-    List<Aula> findByTurmaId(Long turmaId) throws SQLException;
+    void delete(int id) throws SQLException;
+    List<Aula> findByTurmaId(int turmaId) throws SQLException;
 }
